LayoutTests/platform/ios-sim/jquery/offset-expected.txt

ALERT: 383 tests of 387 passed, 4 failed.
